frontserver/requirements.txt
python-daemon
colorama
